fix: bug

+++ Programs/GT9-Team3/Assets/Tower/Scripts/Manager/ResourceManager.cs
@@ -7,7 +7,7 @@ public class ResourceManager : MonoBehaviour
 {
     public static ResourceManager Instance { get; private set; }
 
-    public Dictionary<ResourceType, int> resources = new();
+    public Dictionary<ResourceType, float> resources = new();
 
     void Awake()
     {
@@ -17,28 +17,28 @@ public class ResourceManager : MonoBehaviour
         resources[ResourceType.Tilepiece] = 5000;
     }
 
-    public bool CanAfford(ResourceType type, int cost)
+    public bool CanAfford(ResourceType type, float cost)
     {
         return resources[type] >= cost;
     }
 
-    public void Spend(ResourceType type, int amount)
+    public void Spend(ResourceType type, float amount)
     {
         if (CanAfford(type, amount))
             resources[type] -= amount;
     }
 
-    public void Add(ResourceType type, int amount)
+    public void Add(ResourceType type, float amount)
     {
         resources[type] += amount;
     }
 
-    public int GetAmount(ResourceType type)
+    public float GetAmount(ResourceType type)
     {
         return resources[type];
     }
 
-    public void Earn(ResourceType type, int amount)
+    public void Earn(ResourceType type, float amount)
     {
         if (!resources.ContainsKey(type))
             resources[type] = 0;
@@ -47,12 +47,12 @@ public class ResourceManager : MonoBehaviour
         Debug.Log($"[ÀÚ¿ø] {type} +{amount} È¹µæ, ÇöÀç: {resources[type]}");
     }
 
-    public int ShowTilePiece()
+    public float ShowTilePiece()
     {
         return resources.ContainsKey(ResourceType.Tilepiece) ? resources[ResourceType.Tilepiece] : 0;
     }
 
-    public int showGold()
+    public float showGold()
     {
         return resources.ContainsKey(ResourceType.Gold) ? resources[ResourceType.Gold] : 0;
     }
